Initial commit
none of the systems are linked or active- we still need to narrow in on
which functions lie on which screens, and do some adjustments on the
login functionality when we port it over.  This should just serve as a
beginning layout guide- we may discover we need 1 or 2 more screens to
handle specific game functions.