updated wire framing for gameplay loop
I’ve added “new char” and “resume char” - currently 1 set of player
data is being stored in game preferences, and should probably be
relayed with the server.

Location services are beginning to be present, but they do not seem to
return a value.  I get the debug messages that they are successfully
started, but they don’t seem to get updated values- stay (0,0).  Once
this works “home base” will set a translucent circle on the map.  We
will use map API like google to lay out nearby buildings, with relative
location, and I will work on dummy-ing up an off-state for the
buildings on the map screen in the next build.

Login is still faked. That needs to be done correctly. directly relates
to server interface most likely.

Weapons should carry a durability. may have to create a weapon class to
hold individual stats of same weapon class.  I like the idea of
sticking to “stabbing” “blunt” and “ranged” as sort of the simple core
choice, but then what you’re able to make… pouring your own 50cal, vs
making arrowheads and arrows… determines the stats of that particular
weapon, and the animation that the character.  Likewise the Player
class will have to be built out- where each survivor has their own
stats, so you’re constantly picking the weapon to put in the hands of
which RPG character, and then the deaths will have names attached to
them when you lose people.  Possibly hilarious to use Facebook Login,
and auto-populate from friends list names.