Crit/Miss mechanics added
Right now these are manually wired to the player values, and set by
public variables on the player.  Later they should be calculated as a
function of gear/weapons, but we don’t have the player class for that
yet.

Also added “Drop Loot/ Re-supply” button to map screen. If location
services are on, this will return “confirmed” if the player is within
.00003 deg lat/long from their stored home base location.