Added Food/Water
I added the memory allocation for food and water, and a rough system to
decrease them in 12hr intervals following creation of the character as
the 0 time.  Obviously this will need a lot more work as it just tracks
“meals” and uses total survivors to increment down food and water.

I also made several UI changes- most notably to the start screen.  This
now has some more graphical buttons and layout.  It’s not final, but
it’s clearer and nicer to look at.