mostly graphic updates
I’ve changed a few things with the QR system.  I’m still working
towards the scan creating new survivors on the survivor_roster db, and
then being able to use that survivor immediately.

most of my work needs to be focused on the map screen at this point.
survivors active and alive etc- will be generated based on the status
of the survivors in the list attached to GameManager.instance - in
order for death and reviving to work correctly- this is where they must
be processed and displayed.

also need to work on the home base system.  This should be relatively
easy- to place a green circle on the map in a somewhat accurate way, as
well as store the home base location to the server.