Home base additions
added the home base icon, and the code to place it.  I’ve also added
the ability to update the home base location from the map level via new
menus.  The inventory screen now also includes a “drop gear/resupply”
button that verifies player is in range of home, but does not currently
carry the functions to send the transfer data to the server.