0.59.1 Stable Android Build for Unity 2019.3a
Updated to Unity 2019.3 alpha. I did get this version to output an APK,
but the signing is not working properly with the play store. I’m
committing this so that I can revert the project to an older state,
build out of a non-alpha version of Unity, and see if that fixes it.
This is just a stable version for the Alpha Unity.